feat: add support for Pins API
### Description
This adds functions to add, list and remove pins using the Pins API.

### Regarding `SlackPin` optional message
In the docs for the pin it mentions
> In the past, files and file comments could be pinned to a channel as well.

So I guess this shouldn't happen anymore.

But the following is also documented:

> Different item types can be pinned. Every item in the list has a type property, and the other properties depend on the type of item. The possible types are:
>   message: the item will have a message property containing a message object and a channel property containing the channel ID for the message.
>   file: this item will have a file property containing a file object.
>   file_comment: the item will have a file property containing the file object and a comment property containing the file comment.

When testing for implementing this type, I could not create anything other than `message`.
Files now are just attribute inside a message.

Having the `message` as `Option<SlackMessage>` just protects in case this is used to deserialize
one of these old records. But we could also just have it as `SlackMessage` as should be the case for most now.

### References
- https://api.slack.com/methods/pins.add
- https://api.slack.com/methods/pins.list
- https://api.slack.com/methods/pins.remove